Branch Using the API Version 1
Current code using the old tumblr api (version 1). Contains changes
introduced within the api transition to v2 and the release of TumblThree
v1.0.3.

This contains the user interface changes (save and restore of elements),
proper save of the ratings, tags and notes and possible speed
enhancements due to code cleanup.